// src/findjobnu-api/models/FindjobnuServiceDTOsResponsesJobIndexPostResponse.ts
export interface FindjobnuServiceDTOsResponsesJobIndexPostResponse {
    /**
     * 
     * @type {number}
     * @memberof FindjobnuServiceDTOsResponsesJobIndexPostResponse
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof FindjobnuServiceDTOsResponsesJobIndexPostResponse
     */
    title?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FindjobnuServiceDTOsResponsesJobIndexPostResponse
     */
    company?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FindjobnuServiceDTOsResponsesJobIndexPostResponse
     */
    location?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FindjobnuServiceDTOsResponsesJobIndexPostResponse
     */
    jobUrl?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof FindjobnuServiceDTOsResponsesJobIndexPostResponse
     */
    postedDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof FindjobnuServiceDTOsResponsesJobIndexPostResponse
     */
    category?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FindjobnuServiceDTOsResponsesJobIndexPostResponse
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FindjobnuServiceDTOsResponsesJobIndexPostResponse
     */
    companyUrl?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FindjobnuServiceDTOsResponsesJobIndexPostResponse
     */
    bannerPicture?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FindjobnuServiceDTOsResponsesJobIndexPostResponse
     */
    footerPicture?: string | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof FindjobnuServiceDTOsResponsesJobIndexPostResponse
     */
    keywords?: Array<string> | null;
}